backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt
python-dotenv
requests
twilio
apscheduler
